Treegen + lots of changes

+++ src/main/java/com/uraneptus/lycheed/common/blocks/ModLeavesBlock.java
@@ -23,9 +23,9 @@ public class ModLeavesBlock extends BlueprintLeavesBlock {
         if (rand.nextInt(50) == 0) {
             BlockPos blockpos = pos.below();
             if (world.isEmptyBlock(blockpos)) {
-                double d3 = (double) ((float) pos.getX() + rand.nextFloat());
+                double d3 = (float) pos.getX() + rand.nextFloat();
                 double d4 = (double) pos.getY() - 0.05D;
-                double d6 = (double) ((float) pos.getZ() + rand.nextFloat());
+                double d6 = (float) pos.getZ() + rand.nextFloat();
                 world.addParticle(ModParticleType.LYCHEE_LEAVES_PARTICLE.get(), d3, d4, d6, 0.0D, 0.0D, 0.0D);
             }
         }
